+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Compatibility:
 
 * Fix `Hash#shift` when Hash is empty but has initial default value or initial default proc (@itarato).
 * Make `Array#shuffle` produce the same results as CRuby (@rwstauner).
+* Add `Process.argv0` method (@andrykonchin).
 
 Performance: